Rename Block as it is very specific to the interpreter
