+++ src/som/primitives/bc/block_primitives.py
@@ -1,6 +1,6 @@
 from som.primitives.primitives import Primitives
 from som.vmobjects.primitive   import Primitive
-from som.vmobjects.block       import block_evaluate
+from som.vmobjects.block_bc import block_evaluate, BcBlock
 from som.vm.globals import nilObject, trueObject, falseObject
 
 from rpython.rlib import jit
@@ -25,8 +25,8 @@ jitdriver = jit.JitDriver(
     get_printable_location=get_printable_location)
 
 
-def _execute_block(frame, block, block_method, interpreter, universe):
-    b = universe.new_block(block_method, block.get_context())
+def _execute_block(frame, block, block_method, interpreter):
+    b = BcBlock(block_method, block.get_context())
     frame.push(b)
 
     block_evaluate(b, interpreter, frame)
@@ -37,7 +37,7 @@ def _whileLoop(frame, interpreter, while_type):
     loop_body      = frame.pop()
     loop_condition = frame.pop()
 
-    universe = interpreter.get_universe()
+    # universe = interpreter.get_universe()
 
     method_body = loop_body.get_method()
     method_condition = loop_condition.get_method()
@@ -48,9 +48,9 @@ def _whileLoop(frame, interpreter, while_type):
                                   method_condition=method_condition,
                                   #universe=universe,
                                   while_type=while_type)
-        condition_result = _execute_block(frame, loop_condition, method_condition, interpreter, universe)
+        condition_result = _execute_block(frame, loop_condition, method_condition, interpreter)
         if condition_result is while_type:
-            _execute_block(frame, loop_body, method_body, interpreter, universe)
+            _execute_block(frame, loop_body, method_body, interpreter)
         else:
             break
